refactor : 회사 참여 프로젝트 조회 이동
