+++ Admin/src/main/java/com/seveneleven/service/AdminCompanyService.java
@@ -8,7 +8,6 @@ import java.util.List;
 public interface AdminCompanyService {
     PostCompany.Response createCompany(PostCompany.Request companyRequest);
     GetCompanyDetail.Response getCompanyDetail(Long id);
-    PaginatedResponse<GetProject.Response> getCompanyProject(Integer page, Long id);
     PaginatedResponse<GetCompanies.Response> getListOfCompanies(Integer page);
     PaginatedResponse<GetCompanies.Response> searchCompaniesByName(String name, Integer page);
     PutCompany.Response updateCompany(Long id, PutCompany.Request request);
